@@ -134,6 +134,12 @@ namespace BusinessLibrary.BusinessClasses
                             LoadProperty<int>(RankProperty, reader.GetInt32("RANK"));
                             LoadProperty<int>(TrainingIdProperty, reader.GetInt32("TRAINING_ID"));
                         }
+#if !ER4
+                        if (reader.NextResult())//only happens for "auto reconcile: retain all include codes (raic)"
+                        {//under RAIC, items get unlocked ONLY when user changes item or requests an item (also in "simluate" mode) and get reconciled via the below
+                            ReconcileRAICworker.GetAndDoWork(reader, ri.ReviewId, ri.UserId);
+                        }
+#endif
                     }
                 }
                 using (SqlCommand command2 = new SqlCommand("st_Item", connection))
@@ -164,7 +170,7 @@ namespace BusinessLibrary.BusinessClasses
 
 #endif
 
-    }
+                    }
     [Serializable]
     public class ScreeningItemCriteria: CriteriaBase<ScreeningItemCriteria>
     {
